+++ src/main/php/lang/ast/emit/PHP72.class.php
@@ -13,6 +13,7 @@ class PHP72 extends PHP {
     ArrayUnpackUsingMerge,
     AttributesAsComments,
     CallablesAsClosures,
+    ForeachDestructuringAsStatement,
     MatchAsTernaries,
     NonCapturingCatchVariables,
     NullsafeAsTernaries,